→ Moves: the logic of building the Blogging Prompt card
→ Moves: the logic of building the Blogging Prompt card
to  BloggingPromptCardViewModelSlice
