+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/MySiteViewModel.kt
@@ -40,8 +40,6 @@ import org.wordpress.android.models.JetpackPoweredScreen
 import org.wordpress.android.modules.BG_THREAD
 import org.wordpress.android.modules.UI_THREAD
 import org.wordpress.android.ui.PagePostCreationSourcesDetail.STORY_FROM_MY_SITE
-import org.wordpress.android.ui.bloggingprompts.BloggingPromptsPostTagProvider
-import org.wordpress.android.ui.bloggingprompts.BloggingPromptsSettingsHelper
 import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalOverlayUtil
 import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalOverlayUtil.JetpackFeatureCollectionOverlaySource.FEATURE_CARD
 import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalPhaseHelper
@@ -57,7 +55,6 @@ import org.wordpress.android.ui.mysite.MySiteCardAndItem.Item.SingleActionCard
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.JetpackBadge
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.SiteInfoHeaderCard
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.Type
-import org.wordpress.android.ui.mysite.MySiteCardAndItemBuilderParams.BloggingPromptCardBuilderParams
 import org.wordpress.android.ui.mysite.MySiteCardAndItemBuilderParams.DashboardCardDomainBuilderParams
 import org.wordpress.android.ui.mysite.MySiteCardAndItemBuilderParams.DashboardCardPlansBuilderParams
 import org.wordpress.android.ui.mysite.MySiteCardAndItemBuilderParams.DashboardCardsBuilderParams
@@ -81,7 +78,6 @@ import org.wordpress.android.ui.mysite.cards.DomainRegistrationCardShownTracker
 import org.wordpress.android.ui.mysite.cards.dashboard.CardsTracker
 import org.wordpress.android.ui.mysite.cards.dashboard.activity.ActivityLogCardViewModelSlice
 import org.wordpress.android.ui.mysite.cards.dashboard.bloggingprompts.BloggingPromptCardViewModelSlice
-import org.wordpress.android.ui.mysite.cards.dashboard.bloggingprompts.BloggingPromptsCardAnalyticsTracker
 import org.wordpress.android.ui.mysite.cards.dashboard.domain.DashboardCardDomainUtils
 import org.wordpress.android.ui.mysite.cards.dashboard.domaintransfer.DomainTransferCardViewModel
 import org.wordpress.android.ui.mysite.cards.dashboard.pages.PagesCardViewModelSlice
@@ -127,10 +123,6 @@ import org.wordpress.android.util.SnackbarSequencer
 import org.wordpress.android.util.UriWrapper
 import org.wordpress.android.util.WPMediaUtilsWrapper
 import org.wordpress.android.util.analytics.AnalyticsTrackerWrapper
-import org.wordpress.android.util.config.BloggingPromptsEnhancementsFeatureConfig
-import org.wordpress.android.util.config.BloggingPromptsFeatureConfig
-import org.wordpress.android.util.config.BloggingPromptsListFeatureConfig
-import org.wordpress.android.util.config.BloggingPromptsSocialFeatureConfig
 import org.wordpress.android.util.config.LandOnTheEditorFeatureConfig
 import org.wordpress.android.util.config.MySiteDashboardTabsFeatureConfig
 import org.wordpress.android.util.filter
@@ -144,7 +136,6 @@ import org.wordpress.android.viewmodel.Event
 import org.wordpress.android.viewmodel.ScopedViewModel
 import org.wordpress.android.viewmodel.SingleLiveEvent
 import java.io.File
-import java.util.Date
 import javax.inject.Inject
 import javax.inject.Named
 
@@ -179,7 +170,6 @@ class MySiteViewModel @Inject constructor(
     mySiteDashboardTabsFeatureConfig: MySiteDashboardTabsFeatureConfig,
     private val jetpackBrandingUtils: JetpackBrandingUtils,
     private val appPrefsWrapper: AppPrefsWrapper,
-    private val bloggingPromptsCardAnalyticsTracker: BloggingPromptsCardAnalyticsTracker,
     private val quickStartTracker: QuickStartTracker,
     private val contentMigrationAnalyticsTracker: ContentMigrationAnalyticsTracker,
     private val dispatcher: Dispatcher,
@@ -188,7 +178,6 @@ class MySiteViewModel @Inject constructor(
     private val jetpackFeatureCardShownTracker: JetpackFeatureCardShownTracker,
     private val jetpackFeatureRemovalUtils: JetpackFeatureRemovalOverlayUtil,
     private val jetpackFeatureCardHelper: JetpackFeatureCardHelper,
-    private val bloggingPromptsSettingsHelper: BloggingPromptsSettingsHelper,
     private val jetpackInstallFullPluginCardBuilder: JetpackInstallFullPluginCardBuilder,
     private val bloggingPromptsCardTrackHelper: BloggingPromptsCardTrackHelper,
     private val getShowJetpackFullPluginInstallOnboardingUseCase: GetShowJetpackFullPluginInstallOnboardingUseCase,
@@ -277,7 +266,8 @@ class MySiteViewModel @Inject constructor(
         _onSnackbarMessage,
         siteStoriesHandler.onSnackbar,
         quickStartRepository.onSnackbar,
-        siteItemsViewModelSlice.onSnackbarMessage
+        siteItemsViewModelSlice.onSnackbarMessage,
+        bloggingPromptCardViewModelSlice.onSnackbarMessage
     )
     val onQuickStartMySitePrompts = quickStartRepository.onQuickStartMySitePrompts
     val onTextInputDialogShown = _onTechInputDialogShown as LiveData<Event<TextInputDialogModel>>
@@ -292,6 +282,7 @@ class MySiteViewModel @Inject constructor(
         postsCardViewModelSlice.onNavigation,
         activityLogCardViewModelSlice.onNavigation,
         siteItemsViewModelSlice.onNavigation,
+        bloggingPromptCardViewModelSlice.onNavigation
     )
     val onMediaUpload = _onMediaUpload as LiveData<Event<MediaModel>>
     val onUploadedItem = siteIconUploadHandler.onUploadedItem
@@ -564,14 +555,8 @@ class MySiteViewModel @Inject constructor(
                 postCardBuilderParams = postsCardViewModelSlice.getPostsCardBuilderParams(
                     cardsUpdate?.cards?.firstOrNull { it is PostsCardModel } as? PostsCardModel
                 ),
-                bloggingPromptCardBuilderParams = BloggingPromptCardBuilderParams(
-                    bloggingPrompt = bloggingPromptUpdate?.promptModel,
-                    onShareClick = this::onBloggingPromptShareClick,
-                    onAnswerClick = this::onBloggingPromptAnswerClick,
-                    onSkipClick = this::onBloggingPromptSkipClick,
-                    onViewMoreClick = this::onBloggingPromptViewMoreClick,
-                    onViewAnswersClick = this::onBloggingPromptViewAnswersClick,
-                    onRemoveClick = this::onBloggingPromptRemoveClick
+                bloggingPromptCardBuilderParams = bloggingPromptCardViewModelSlice.getBuilderParams(
+                    bloggingPromptUpdate
                 ),
                 domainTransferCardBuilderParams = domainTransferCardViewModel.buildDomainTransferCardParams(
                     site,
@@ -1240,66 +1225,6 @@ class MySiteViewModel @Inject constructor(
         mySiteSourceManager.refresh()
     }
 
-    private fun onBloggingPromptShareClick(message: String) {
-        _onNavigation.value = Event(BloggingPromptCardNavigationAction.SharePrompt(message))
-    }
-
-    private fun onBloggingPromptAnswerClick(promptId: Int) {
-        bloggingPromptsCardAnalyticsTracker.trackMySiteCardAnswerPromptClicked()
-        val selectedSite = requireNotNull(selectedSiteRepository.getSelectedSite())
-        _onNavigation.value = Event(BloggingPromptCardNavigationAction.AnswerPrompt(selectedSite, promptId))
-    }
-
-    private fun onBloggingPromptSkipClick() {
-        selectedSiteRepository.getSelectedSite()?.let { site ->
-            val siteId = site.localId().value
-
-            appPrefsWrapper.setSkippedPromptDay(Date(), siteId)
-            mySiteSourceManager.refreshBloggingPrompts(true)
-
-            val snackbar = SnackbarMessageHolder(
-                message = UiStringRes(R.string.my_site_blogging_prompt_card_skipped_snackbar),
-                buttonTitle = UiStringRes(R.string.undo),
-                buttonAction = {
-                    bloggingPromptsCardAnalyticsTracker.trackMySiteCardSkipThisPromptUndoClicked()
-                    appPrefsWrapper.setSkippedPromptDay(null, siteId)
-                    mySiteSourceManager.refreshBloggingPrompts(true)
-                },
-                isImportant = true
-            )
-
-            _onSnackbarMessage.value = Event(snackbar)
-        }
-    }
-
-    private fun onBloggingPromptViewAnswersClick(promptId: Int) {
-        bloggingPromptsCardAnalyticsTracker.trackMySiteCardViewAnswersClicked()
-        val tag = BloggingPromptsPostTagProvider.promptIdSearchReaderTag(promptId)
-        _onNavigation.value = Event(BloggingPromptCardNavigationAction.ViewAnswers(tag))
-    }
-
-    private fun onBloggingPromptViewMoreClick() {
-        _onNavigation.value = Event(BloggingPromptCardNavigationAction.ViewMore)
-    }
-
-    private fun onBloggingPromptRemoveClick() {
-        launch {
-            updatePromptsCardEnabled(isEnabled = false).join()
-            _onNavigation.value = Event(BloggingPromptCardNavigationAction.CardRemoved)
-        }
-    }
-
-    fun onBloggingPromptUndoClick() {
-        bloggingPromptsCardAnalyticsTracker.trackMySiteCardRemoveFromDashboardUndoClicked()
-        updatePromptsCardEnabled(true)
-    }
-
-    private fun updatePromptsCardEnabled(isEnabled: Boolean) = launch {
-        selectedSiteRepository.getSelectedSite()?.localId()?.value?.let { siteId ->
-            bloggingPromptsSettingsHelper.updatePromptsCardEnabled(siteId, isEnabled)
-            mySiteSourceManager.refreshBloggingPrompts(true)
-        }
-    }
 
     private fun onJetpackFeatureCardClick() {
         jetpackFeatureCardHelper.track(Stat.REMOVE_FEATURE_CARD_TAPPED)
